requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6

# Selenium e dependências
selenium==4.25.0

# Outras dependências
requests
psycopg2-binary
python-dotenv
tabulate